aarch64: access stack in pairs of registers
On aarch64 CPUs load and store addresses have to be aligned to multiples
of 16. Otherwise an exception is thrown and the application receives a
SIGBUS. Unlike actual CPUs, QEMU does not care about alignment here.

Use stp and ldp instructions for manipulating the stack in register
pairs which automaticlly leads to appropriately aligned addresses.

Signed-off-by: Josua Mayer <josua@solid-run.com>